refactor: extract request authority mapping
