+++ crates/coven-cli/src/main.rs
@@ -20,6 +20,7 @@ use uuid::Uuid;
 mod afs;
 mod afs_mount;
 mod api;
+mod api_health;
 mod api_response;
 mod api_routes;
 mod capabilities;
@@ -67,6 +68,7 @@ mod proposal_scheduler;
 mod proposal_store;
 mod pty_runner;
 mod repos_config;
+mod request_authority;
 mod reset;
 mod session_launch;
 mod settings;
